valoranterus

+++ src/cogs/etc/presets.py
@@ -126,7 +126,7 @@ def add_points(user, cur, payload):
     if not isinstance(cur, mysql.connector.cursor.MySQLCursor):
         raise ProgrammingError('Cur Argument is not an MySQLCursor Object')
 
-    current_exp = payload[1] + (2 * payload[2])  # EXP Addition
+    current_exp = payload[1] + int((2 * float(payload[2])))  # EXP Addition
 
     cur.execute("UPDATE points SET Experience=%s WHERE User=%s;", (current_exp, user))
     dbBase.commit()
